Quick Edits

+++ setup.sh
@@ -13,13 +13,28 @@ TAGS_DIR="./wd14_tagger_model"
 mkdir -p "$MODEL_DIR"
 mkdir -p "$TAGS_DIR"
 
-# Download the model file
-echo "Downloading ONNX model..."
-curl -L "https://huggingface.co/$REPO_ID/resolve/main/$MODEL_FILE" -o "$MODEL_DIR/$MODEL_FILE"
+echo
+echo "Gathering asset files"
+echo
+# Check if the model file already exists
+if [ -f "$MODEL_DIR/$MODEL_FILE" ]; then
+  echo "ONNX model already exists. Skipping download."
+else
+  # Download the model file
+  echo "Downloading ONNX model..."
+  curl -L "https://huggingface.co/$REPO_ID/resolve/main/$MODEL_FILE" -o "$MODEL_DIR/$MODEL_FILE"
+fi
 
-# Download the tags file
-echo "Downloading tag list..."
-curl -L "https://huggingface.co/$REPO_ID/resolve/main/$TAGS_FILE" -o "$TAGS_DIR/$TAGS_FILE"
+# Check if the tags file already exists
+if [ -f "$TAGS_DIR/$TAGS_FILE" ]; then
+  echo "Tag list already exists. Skipping download."
+else
+  # Download the tags file
+  echo "Downloading tag list..."
+  curl -L "https://huggingface.co/$REPO_ID/resolve/main/$TAGS_FILE" -o "$TAGS_DIR/$TAGS_FILE"
+fi
+
+echo
 
 # Create and activate a Python virtual environment
 echo "Creating and activating virtual environment..."
@@ -29,5 +44,8 @@ source venv/bin/activate
 # Install dependencies from requirements.txt
 echo "Installing dependencies..."
 pip install -r requirements.txt
-
+echo
 echo "Setup complete!"
+echo
+echo "You may run: python start.py"
+echo
